attempt at refactor (broken)
