+++ packages/editor/src/state/atoms.ts
@@ -6,7 +6,6 @@ import {
 } from "@penrose/core";
 import { PathResolver, Trio, TrioMeta } from "@penrose/examples/dist/index.js";
 import registry from "@penrose/examples/dist/registry.js";
-import { User as FirebaseUser } from "firebase/auth";
 import { Actions, BorderNode, TabNode } from "flexlayout-react";
 import localforage from "localforage";
 import { debounce, range } from "lodash";
@@ -106,7 +105,7 @@ export type AuthModalState = {
   registerIsOpen: boolean;
 };
 
-export type AppUser = FirebaseUser | null;
+export type AppUser = boolean;
 
 const localFilesEffect: AtomEffect<LocalWorkspaces> = ({ setSelf, onSet }) => {
   setSelf(
@@ -242,8 +241,8 @@ export const currentAuthModalState = atom<AuthModalState>({
 });
 
 export const currentAppUser = atom<AppUser>({
-  key: "currentAppuser",
-  default: null,
+  key: "currentAppUser",
+  default: false,
 });
 
 /**
